05 - OpenAI code interpreter, chat with data/requirements.txt
openai==1.3.0
langchain==0.0.267
langchain_community==0.0.10
faiss-cpu==1.7.4
streamlit==1.28.0
python-dotenv==1.0.0
pandas==2.0.3
matplotlib==3.7.2
requests==2.31.0
pypdf==3.17.0
tiktoken==0.5.2
